extract tree to submodule
